ASC-127 Move derived groups to the parameters folder
