+++ atlast_sc/instruments/classes/Default.py
@@ -1,7 +1,7 @@
 import astropy.units as u
 from astropy import constants
 from atlast_sc.core.instrument import Instrument
-from atlast_sc.derived_groups import noise_temperature
+from atlast_sc.parameters.derived_groups import noise_temperature
 
 """
 Default instrument parameters
