+++ atlast_sc/instruments/classes/Finer.py
@@ -1,7 +1,7 @@
 import re
 import astropy.units as u
 from atlast_sc.core.instrument import Instrument
-from atlast_sc.derived_groups import noise_temperature
+from atlast_sc.parameters.derived_groups import noise_temperature
 
 """
 FINER instrument parameters
