+++ atlast_sc/instruments/classes/Tifuun.py
@@ -2,7 +2,7 @@ from numpy import expm1, sqrt
 import astropy.units as u
 from astropy import constants
 from atlast_sc.core.instrument import Instrument
-from atlast_sc.derived_groups import noise_temperature
+from atlast_sc.parameters.derived_groups import noise_temperature
 
 """
 TIFUUN instrument parameters
